restructured component tree and renamed components
