+++ client/src/components/todo/styles.js
@@ -1,0 +1,12 @@
+const styles = theme => ({
+  toolbar: {
+    display: 'flex',
+    justifyContent: 'space-between',
+  },
+  link: {
+    textDecoration: 'none',
+    color: 'inherit',
+  },
+});
+
+export default styles;
