+++ client/src/components/todoDashPage/styles.js
@@ -1,8 +1,4 @@
 const styles = theme => ({
-  toolbar: {
-    display: 'flex',
-    justifyContent: 'space-between',
-  },
   loading: {
     marginTop: '50%',
     padding: theme.spacing.unit * 3,
@@ -30,10 +26,6 @@ const styles = theme => ({
   },
   text: {
     textDecoration: 'none',
-  },
-  link: {
-    textDecoration: 'none',
-    color: 'inherit',
   },
 });
