+++ client/src/components/viewTodoPage/styles.js
@@ -1,8 +1,4 @@
 const styles = theme => ({
-  toolbar: {
-    display: 'flex',
-    justifyContent: 'space-between',
-  },
   loading: {
     marginTop: '50%',
     padding: theme.spacing.unit * 3,
@@ -12,17 +8,6 @@ const styles = theme => ({
     [theme.breakpoints.up('lg')]: {
       marginTop: '20%',
     },
-  },
-  title: {
-    width: '100%',
-    textAlign: 'center',
-    padding: theme.spacing.unit * 3,
-  },
-  main: {
-    display: 'flex',
-    justifyContent: 'center',
-    width: '100%',
-    marginTop: '50px',
   },
   paper: {
     width: '90%',
@@ -37,19 +22,28 @@ const styles = theme => ({
       width: '40%',
     },
   },
-  form: {
-    width: '60%',
-    margin: '0 auto',
+  main: {
+    display: 'flex',
+    justifyContent: 'center',
+    width: '100%',
+    marginTop: '25px',
+  },
+  navIcons: {
+    display: 'flex',
+    justifyContent: 'flex-start',
+    width: '100%',
+    marginTop: '25px',
+    paddingLeft: '50px',
   },
   buttonWrapper: {
     display: 'flex',
-    width: '100%',
-    justifyContent: 'space-around',
+    justifyContent: 'flex-end',
+  },
+  title: {
     padding: theme.spacing.unit * 3,
   },
-  link: {
-    textDecoration: 'none',
-    color: 'inherit',
+  description: {
+    padding: theme.spacing.unit * 3,
   },
 });
